set nox min version only for newer sys python
The min version here fails on github actions because the runners
there use the single version of python for the install, and the newer
nox is not available on older pythons like 3.7.

We probably dont need the min version anyway, as since we have just
switched to venv in 5e62ea65167c, the misinterpretation of
"python3.14t" doesn't seem to happen anyway.  however, all versions
of nox seem willing to install a 3.14t interpreter when 3.14 is
requested, and a 3.14t is sooner in the path.  does not seem to
happen the other way around though (e.g. when 3.14t is requested,
it does not use a 3.14)

Anyway, for better predictability leave a min nox version in place
where we can do so for now.

Change-Id: I93a2c2e71798f5a75d71cc63de54750ec13975a1
(cherry picked from commit 93968e2c9e0796a4ea2fc13582b2eec83371f008)